pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "debass_meta"
version = "0.1.0"
description = "DEBASS meta-classifier: trust-aware fusion of Rubin broker outputs"
requires-python = ">=3.10"
dependencies = [
    "alerce>=2.3.0,<3",
    "fink-client>=7.0",
    "lasair>=0.1.0,<1",
    "lightgbm>=4.0",
    "scikit-learn>=1.4",
    "pandas>=2.0",
    "pyarrow>=15.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
local-experts = [
    "supernnova>=2.0",
    "astro-parsnip>=1.4.0",
    "astropy>=5.0",
    "lcdata>=0.3",
]
dev = [
    "pytest>=8.0",
    "pytest-cov",
]

[project.scripts]
debass_meta-probe = "debass_meta.cli.probe:main"
debass_meta-backfill = "debass_meta.cli.backfill:main"

[tool.hatch.build.targets.wheel]
packages = ["src/debass_meta"]

[tool.pytest.ini_options]
testpaths = ["tests"]
